# service/openapi.yaml
openapi: 3.0.3
info:
  title: Temis DMN Engine API
  version: "1.0-experimental"
  description: >
    HTTP interface to the Temis DMN 1.5 decision engine. Upload a DMN XML model,
    then evaluate one of its decisions against a JSON input context. A stateless
    endpoint compiles and evaluates in a single request. The surface is
    experimental until the v1 stabilisation (roadmap WP-43).


    When the server is started with a token (`temisd -token …` or
    `TEMIS_API_TOKEN`), the `/v1` endpoints require an
    `Authorization: Bearer <token>` header — click **Authorize** and paste the
    token to try them here. When no token is configured the API is open and the
    header is ignored.
servers:
  - url: /
# Bearer auth is optional: when the server runs without a token the empty
# requirement ({}) makes every endpoint callable; when a token is set, callers
# present it via the bearerAuth scheme. Swagger UI's "Authorize" button binds to
# bearerAuth.
security:
  - bearerAuth: []
  - {}
paths:
  /v1/models:
    post:
      summary: Compile and cache a DMN model
      description: >
        The request body is the raw DMN XML document. The response carries the
        content-addressed model id, the model index and any compile diagnostics.
        Re-uploading the same document returns the same id.
      requestBody:
        required: true
        content:
          application/xml:
            schema: { type: string }
      responses:
        "201":
          description: Model compiled and cached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelResponse" }
        "400":
          description: Empty body or malformed XML
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      summary: List cached models
      description: >
        Returns a summary (id, decisions, inputs) of every model currently held
        in the in-memory cache. This endpoint can be disabled at startup
        (`temisd -list-models=false`) to keep the cached decisions private; when
        disabled it responds 404 as if it did not exist.
      responses:
        "200":
          description: The cached models
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelListResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Model listing is disabled
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/models/{id}:
    get:
      summary: Get a cached model's index
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Model index
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No model with that id
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/models/{id}/evaluate:
    post:
      summary: Evaluate a decision of a cached model
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluateModelRequest" }
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluateResponse" }
        "400":
          description: Invalid request body or missing decision
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown model or decision
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: Evaluation failed (e.g. UNIQUE table with multiple matches)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/evaluate:
    post:
      summary: Stateless compile-and-evaluate in one request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluateStatelessRequest" }
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluateResponse" }
        "400":
          description: Missing xml/decision or malformed input
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /docs:
    get:
      summary: Interactive Swagger UI
      description: Browser page for trying the API; loads the spec from /openapi.yaml.
      security: []
      responses:
        "200": { description: HTML page }
  /openapi.yaml:
    get:
      summary: This OpenAPI document
      security: []
      responses:
        "200": { description: OpenAPI 3 document (YAML) }
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "200": { description: OK }
  /readyz:
    get:
      summary: Readiness probe
      responses:
        "200": { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Opaque token configured via `temisd -token` / `TEMIS_API_TOKEN`. Sent as
        `Authorization: Bearer <token>`. Only enforced when the server is started
        with a token.
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when the server requires one)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    ModelListResponse:
      type: object
      properties:
        models:
          type: array
          items: { $ref: "#/components/schemas/ModelSummary" }
        count: { type: integer }
    ModelSummary:
      type: object
      properties:
        modelId: { type: string, example: "sha256:..." }
        decisions:
          type: array
          items: { type: string }
        inputs:
          type: array
          items: { type: string }
    ModelResponse:
      type: object
      properties:
        modelId: { type: string, example: "sha256:..." }
        decisions:
          type: array
          items: { type: string }
        inputs:
          type: array
          items: { type: string }
        diagnostics:
          type: array
          items: { $ref: "#/components/schemas/Diagnostic" }
    EvaluateModelRequest:
      type: object
      required: [decision]
      properties:
        decision: { type: string }
        input:
          type: object
          additionalProperties: true
    EvaluateStatelessRequest:
      type: object
      required: [xml, decision]
      properties:
        xml: { type: string }
        decision: { type: string }
        input:
          type: object
          additionalProperties: true
    EvaluateResponse:
      type: object
      properties:
        outputs:
          type: object
          additionalProperties: true
        decisions:
          type: object
          additionalProperties: true
        diagnostics:
          type: array
          items: { $ref: "#/components/schemas/Diagnostic" }
    Diagnostic:
      type: object
      properties:
        severity: { type: string, enum: [error, warning, info] }
        code: { type: string }
        message: { type: string }
        decisionId: { type: string }
        line: { type: integer }
        col: { type: integer }
    Problem:
      description: RFC-7807 problem detail with a stable engine-specific code.
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }